Create a Subagent in mib_module to handle GET and SET request on deviceType
To perform set and get request on the VIVOE-MIB:
    1/ run the snmp deamon with:
"sudo snmpd -f -Lo -C --rwcommunity=public --master=agentx
--agentXSocket=tcp:localhost:705 udp:161"
this runs the deamon on port 161, the deamon run the Xagent
on local port 705. When receiving a SET or GET request, snmpd
transmits the request to the Xagent, that query the MIB through
the subAgent I have written, and handle the request.
    2/ on the agent, run the subAgent (the one who handles the
    request"
"./deviceType -f -Lo -x  tcp:localhost:705"
Add -DALL option in both commands to add the debug trace.