mib_module: adding MIB parameters (all of the deviceInfo section unless the table)
I finnished to add all the parameters of the deviceInfo section of the MIB, unless
the  ethernetIfTable. I need to study the net-snmp documentation to understand how
table works. Otherwise I have the subAgent and it works fine with both Read-Only and
Read-Write information. After implementing the Table, I will work on the initiation
of the subAgent through a "*.conf" file.