mib_module: creation of the part of the subAgent the handle the table: ethernetIfTable
Actually, this has revealed to be a bit disappointing. As the code written previously, has revealed
to be useless. Indeed the file mib_module/src/ethernetIfTable.c does now contain the code that
hanldes the GET, GETNEXT and SET requests of the objects that are in that table:
_ethernetIfIndex
_ethernetIfSpeed
_ethernetIfMacAddress
_ethernetIfIpAddress
_ethernetIfSubnetMask
_ethernetIfIpAddressConflict
So I delete the files which are no more usefull.
Furthermore, the object ethernetIfEntry of the VIVOE-MIB, does not need to be implemented in the
same way as the other objects described above. This is why the access rights to this object are
set to "not accessible" in the MIB. This object represents a whole line of ethernetIfTable. It is
a sequence of the objects described above. In the code of the subAgent, in the file ethernetIfTable.c
it is represented as a structure that includes those objects. But, once again, there is no need to implement
it the usual way.
Now that we handle the GET, GETNEXT and SET requests for all the objects in the MIB that are part of the
device section; the next step will be to use the glibc to handle the initialization of those parameters
from the command line or a config file.