mib_module: changes in hte way to manage IP addresses, improvement in the initialization of the mib through vivoe-mib.conf
The old way to manage IP was to use string to represent them. However, there is a type to represent
IP addresses in the kernel named in_addr_t which is a 32-bits integer. It is only the decimal value
of an IP address reprensented as a binary string. The new files here used this representation
(especially ethernetIfTable.c).
There are also improvement in the way of initialize the MIB through the configuration file. Indeed,
now, a difference is made between optional parameters as deviceUserDesc and compulsory parameters
as ethernetIfNumber. The initialization of the MIB, and the program itself stop when a compulsory
parameter is not initialized. Also, there are a few changes with the parameters of the ethernetIfTable,
as there can be several of them (one for each line of the table), so there are initialized into arrays
of different types. In the vivoe-mib.conf file they are represented as lists.
Only thing to improve now, is the handling of ethernetIfMacAddress as an array of byte.