mib_module: improve the robustness in the way to initialize the Mac Address in the MIB
Previously a Mac address could be initialized through the vivoe-mib.conf file to
any string value. Now, the string need to have the format:
XX:XX:XX:XX:XX:XX
Where X is a character included between 0-->9, a--f, or A-->F.
The length need to be exaclty 17 character (excluding '\0').
The parser between to byte's value should be exactly ':' and no other separator.
The string is then converted into an array of byte directly during the initialization
of the entries the ethernetIfTable.