mib_module: cleaning the code: creation of structures to handle mib parameters
Before there was only one structure that handled the MIB's parameters.
It was just a structure that contained parameters of different types(int, char*
int* or char**), so their initialization should be done individually, there was
no parallelizationof the code, which was not portable nor scalable. This comit
fixes this issue by creating two different structures: one to represents a
MIB's parameter such as: deviceDesc, deviceMIBversion, ethernetIfnumber or
channelStatus; and another one to represents a MIB's group such as deviceInfo,
channelControl or vivoeNotifications. The second structure contains a array of
the first one.
This is the initialization of the MIB's parameters is parallelized and scalable.