streaming: removing gst_parse_launch from the code
Previously, gst_parse_launch wa used to create a GstElement with
rtppay and udpsink already linked into it. Now we are creating two
different elements, one for the RTP payload encryption, and another one
for the creation of the udpsink.
Then we link them both manually to the other GstElement of the pipeline.