streaming: creating SDP files to decode the stream
Before, the command line use to decode the stream used udpsrc to receive
the stream, passing the "caps" to the command line. This is no longer
needed, form the caps we create a SDP file, (for Session Description
Protocol) which described the UDP source stream. This file is pipeline
with the plug-in sdpdemux (plug-in use to read the SDP file inside
GStreamer) and then we proceed to the classic pipeline to decode and
play the stream.