streaming: reaoganize functions, define filter
This commit reorganize the code into the different functions for the
wright files. It also defines the filter for raw video formats (they are
obtained from the command gst-launch-1.0 rtpvrawpay) because all video
format for videotestsrc are RGB, YCbCr or Gray (monochrome), but not all
of them can be used as source for rtpvrawpay. So we limit the video
formats to use to RBG, YUV and Gray format supported by rtpvrawpay.