streaming: put mpeg-4 encoder into main function
Before the MPEG-4 encoder was created and linked to the pipeline at the
same time as the rtpmp4vpay and the udpsink were. However, this do not
represent the reality. Indeed, MPEG-4 videos will be the output
video format of our camera when we will use cameras and no videotestsrc.
So the encryption do not need to be perform when we create the pipeline
for MPEG-4 video but when we create a fake source, so in the main
function.