streaming:Add detection of input video source caps
First idea was to detect the input video format: RAW or MPEG-4, to use
the correct pipeline and the correct filter to filter out non VIVOE
video formats. But, as we will need all input source video capabilities
to initiate the MIB it have been done at the same time. So this commit
add into main.c functions that are used to detect all the input source video
capabilities (format, encoding, resolution, etc...) just before link it
in the pipeline to the RTP payloader element.