openvivoe: merge streaming and mib_module
Streaming was the project (binary, sources, ressources) that handeled the
streaming part of the openvivoe project, independently of mib_module
which was the deamon of the subAgent which handeled the MIB managament
part of the openvivoe project. Now those two parts of the project have
been merge into one single program as we aim to initialize the MIB with
information extract from Gstreamer.