streaming: start to intiate the MIB from Gstreamer values
In stream.c a function has been created to intiate the MIB from the
capabilies of the input source video detected. But first we need to make
sure that all MIB's paramaters of videoFormatInfo can be initialize from
the caps detected. So we have two options: we can try to detect
automatically their values from the caps we already knows, or we can
impose to the user to specify them all.