streaming: create detect.c to handle capabilities detection
Before the dectection capabilities was done in stream.c, but it occurs
that we need to do it several times: once in stream.c before linking the
source to the rtp payloader. And once after linking the source and the
payloader, but before linking the rtp payloader to the udpsink.