Add the initiation of videoFormatTable from gstreamer information
The initiation of the table of videoFormatTable is done directly from
the stream, as we are able to get the caps of the stream. A maximum of
parameters are initiated, but not all parameters (for now, the case of
ROI has not been taken into account). The group videoFormatInfo of the
MIB is not defined nor initialized the same way than deviceInfo. Indeed,
as it only contains a parameter and a table, no MIB_Group global var is
defined. But the entry in the table are created dynamically when a new
videoFormat is stream on the device. The parameter videoFormatNumber is
automatically updated. When a stream is launched but with a videoFormat
already present in the table, the format is not added but we only set
its "videoFormatStatus" to "enable (1)". However, this requires to set
the "videoFormatStatus" parameter to "disable (0)" each time a stream is
stopped in Gstreamer.
 However, this functionnality has
not been tested yet, but it will be done in the next commit.