videoFormatTable: add function to initiate entry
Before, the only function that existed was videoFormatTableEntry_create
which created an entry with parameters specified and linked it into the
list of entries of the table. Now the function create uses the function
which only initiates the entry (set the value to parameters but does not
link it in the list), and then link the entry in the list.
This difference between those to function is made to be able to initiate
an entry in several times. Indeed we want to initiate a maximum of
parameters of the entry from the inforamtion of Gstreamer, so we will
add information to the entry each time we add an element to our
Gstreamer pipeline.