Initiate an videoFormatTableEntry form Gstreamer
The initiation of the parameters of videoFormatTable_entry works. For
now, each time we add an element to the pipeline wich is not a sink,
these functions should be performed, as the video format may be changed.
The only things remaining now is to add this entry into the table. But
before, we should make sure that this videoFormat is not already in the
table. The function already exists, but is not up-to-date, as some
features in the code changed since it had been declared.