Tests that the initiation of videoFormatTable works
From this commit the initiation of the table for an empty table, and of
a table that already contains an entry works.
Final test to do: check that if an entry already exists in the table,
it is not added to the table a second time, and that its status is set
to enable.