stream_registration: add robustness to registration
Now we check if videoFormatNumber equals 255 before adding a new entry
in the table. If it is 255 we do not add a new entry in the table.
In the conditions to wich are the parameters of the caps of the video
stream, I removed second part (aftre && conditions).
Indeed this part was used to check if the entry is an "empty entry" but
the concept of "empty entries" has been removed from the code, so this
code was no longer needed, and could provide errors.