streaming: possibility to limits VIVOE formats
Using the configuration file vivoe-stream.conf, the user has now the
possibility to limit the VIVOE formats he wants to use. For example he
can specify through the configuration file that he only wants his VIVOE
to work with MPEG-4 videos.
Furthermore, the way to build the filter that filters out non VIVOE
video formats has changed. It is more efficient now, as we do not need
to specify, when we are building the pipepline if we wanna use the
MPEG-4 filter or the RAW filter. The filter is now built to allow both
format in only one caps. Besides, as the user can aslo choose through
the configuration file which resolutions, and encodings he wants. The
caps for the filter are not buit from macros anymore, but built
dynamically directly into the code. It decreases performances, but
scalabilty of functionnalities and of the code is greatly improved.