Add the computation of multicast IP addresses
The norm VIVOE gives the possibilit to stream to multicast addresses in
the range: 239.192.1.1 to 239.192.254.254. Thanks to the code of Gerome
Burlats <gerome.burlats@openwide.fr> I have added the possibility to
compute the VIVOE multicast IP address on which the stream will be sent.
To compute this address we retrieve the IP address of the interface that
we want to use for the VIVOE protocol, from this IP address, the last
byte is extracted, and it defines the deviceID. Then we just need to
concatenate "239.192" + channelID + deviceID where channelID is the
channel ID we want to use for our stream.