Adding files used to retrieve internet interface info
From the code generate in previous commit. We extend this code to get
not only the IP address of the ethernet interface specified, but also,
its ethernet MAC address, the subnet Mask used and the Speed or MTU
(Maximum Transfer Unit). This values will then be used to initiate the
MIB, and the configuration file vivoe-mib.conf will no longer be nedeed
for this part of the MIB.