Correcting a bug in ioctl calls
In the file ethernetIfTabke_init.c we initiate the MIB directly from
information retrieved from the ethernet Interface. But the structure that
we pass to ioctl ifreq register the information received by ioctl in a
union. So each call to ioctl erase the result from a previous call. This
was what happened.