Get network interfaces from vivoe-mib.conf or user input
Before, the vivoe-mib.conf aimed to initialize all the MIB with
parameters entered by the user. One of the improvement here is to use
ioctl calls to retrieve information on the network interfaces that are
using VIVOE. However, the user still need to specify which NI he wants
VIVOE to use. Now the user have the availability to specify them through
the vivoe-mib.conf or through the standard input when the program is
running. Next steps are to correct a bug in the Ip address and the
SubnetMask retrieved. And then, to detect that the user has specified
the NI to use in vivoe-mib.conf to prevent from asking him twice.