Correct bug in the intiation of IP and Mask of ethernetIfTable
Before we were using inet_netof to convert an in_addr to an unsigned
long int, but it is not what it does. Indeed, according to the man
page, it returns the net_id of in_addr. So we simply get the
parameter s_addr of in_addr to initialize the parameters
ethernetIfIpAddress and ethernetSubnetMask. Now we have succeed to
initialize the Table ethernetIfTable dynamically, using the
configuration file vivoe-mib.conf to get the network interface(s)
that will use VIVOE.