Add features to stop stream nicely
Before, the stream was run in the GMainLoop and did not stop the
GMainLoop when a SIGINT or SIGTERM was send to the program. So the
problem was that if you try running two threads, one for the stream one
for the subAgent of the MIB. Exiting the program with Ctrl-C was not
possible as only the MIB's SubAgent was stopped. Now both the streaming
part and the MIB's Subagent handle the signals SIGINT and SIGTERM.
Next step is to run them in different threads.