Run the snmp request handler periodically inside the GMainLoop
Before, the streaming part and the deamon of the SubAgent that handles
the SNMP requests could not run at the same time. Now, they can. This
commit add to the MainLoop an idle task, that is called every 100ms.
This task calls a function that handles the SNMP requests if there are
pending ones. When SIGINT or SIGTERM is received, it stop the deamon,
then exit the MainLoop.