Allows to start and stop pipeline as you want
Before, the initation of the pipeline and the starting of the stream
were done in hte same function. Now, there are 4 different functions
to handle the stream:
_ init_stream
_ start_stream
_ stop_stream
_ delete_stream_datas

The first one is used to build the pipeline: add the elements and link
them. The second is used to set the pipeline state to PLAY. The third is
used to set the pipeline state to stop. The fourth to delete the
pipeline, and the bus_watch_id associated to it.