Add channelControl support in snmp subAgent deamon
This commit add the support of a new group of the MIB: channel Control.
This group describes a video format that is streamed on a multicast IP.
It contains three parameters channelNumber, the number of channels,
channelReset a parameter used to reset one or several channels, and a
channelTable which describe the channels.