Modify the way to initiate entries in MIB tables
Before, to add an entry that has a string parameter, two  kinds of
parameters were passed to the function: the string and its length.
However, our MIB defines for each string a maximum length which is 16,
32 or 64 bytes. So now, only the string is passed as a parameter to the
function create_entry, and the length is dynamically computed as the
minimum between [16,32,64] and the actual string's length.