Save videoFormatIndex in stream_data when adding a new format in table
When adding a new videoFormat in the videoFormat Table, we save
the index inside the structure that represents the data associated to
the stream. This will be useful in several cases:
_when several streams are launched, to know for which entry
in the table we must set the parameter "videoFormatStatus" to disable.
 _when we want to initialize a new entry in channelControl, this will be
used to retrieve all the parameters we need to intialize it.