Add the initiation of channelControl
When we fill the videoFormatTable with information extract from
Gstreamer, we also intiate the channelControl by creating a new entry
containing the values that need to be intiate from videoFormat. Litle
bug to correct: snmpget on Ip addresses, and length of string parameters
for channelControl.
Next step add start_stream function in the handler of channelStatus to
launch the stream through snmp requests.