Fixes in the way to create entries in channelTable
There were several problems: the multicast address were calculated in
network mode but should be calculated in host mode as the deamon
transforms them when snmp request for ip addresses is done. Ther way to
answer to snmp request for IP addresses was not good. All of this is
fixed, channelTable is now initialized properly, and channelTable anwsers
to snmp GET methods correctly.