Get multicast IP address for stream from MIB parameters
Before, IP address for the udpsink was passed to parameters of the
program. But now it is directly computed using multicast.c and
define_vivoe_multicast from the videoFormatIndex we are registering and
the IP of the network interface used. This computed IP is used to
initialize the channelTable and the "ip" parameter of gstreamer's
element udpsink.