Check that create, start and stop several streams works
This commit fixes and tests that we have the behaviour we want for the
Service Provider Part of the VIVOE norm. Thus, we can launch several
streams, and register them in videoForamtTable if their different in
some way (resolution, encoding or whatever). We can play them, starting
them by setting their ChannelStatus to 'start' the channelTable. We can
stop them the same way. This streams used the multicast IP computed and
stored in the MIB (under the parameter channelReceiveIpAddress in the
channelTable).