Fixes in the way stop stop and exit program
The program catches signals SIGINT and SIGTERM to exit properly. When
exiting the program, each stream is stopped, and all data relative to
this stream are deleted (freed). Before, the stop_data contained a member
which was the stream that has been created. But as now several streams
can be created, stop_data does no longer contain a stream but an array
of streams to stop and delete them all when exiting the program.