Change the way to build filter for VIVOE video formats
The way to build the filter before did not work for MPEG-4 videos.
Indeed, I was using gst_link_filtered to filters out non VIVOE formats.
However, the function just build a capsfilter and tries to link it with
the two Gstreamer's elements to link. So in the case of RAW videos it
tries to link a capsfilter to another with two different resolutions,
the filter worked. But in reality their will be no capsfilter in the
pipeline, and our link will work, as we have no possiblity to know the
resolution of the input video without setting pipeline in playing
state. This was what happened for the MPEG-4 videos. The solution
implemented here is to get the caps of the input video and the try
intersect them with the caps of our VIVOE filter. If the intersection is
empty, the videoFormat is filtered out, otherwise it is not, and a link
is made between the two Gstreamer's elements.