Code Cleaning: improve performance in handler of channelTable
Before, in the handler of channelTable in channelTable.c there were two
verifications that were done on the values sent by the user for writable
lines in the Table: first one was type checnking done the cas
'SET_RESERVE_1', second was value checking: for example check that int
value gave by the user is not negative for resolution, this verification
was done in 'SET_RESERVE_2'. Now both of those verifications are done at
the same time in SET_RESERVE_1.