Change the way to manage channels: use the correct way!
Before, each time a stream was initialized, a videoFormatEntry AND a
channelEntry were created and added into their respective tables. So at
the end, there were as many channels as videoFormats, so the mapping of
a videoFormat to a channel was meaningless.
However, this is not how VIVOE should work. There should not be as many
channels as videoFormats, but as many channels as network interfaces.
Used for VIVOE.
This is why a map is established by the manager between a videoFormat
and a channel, because, most of the time (and I'm talking for something
like 99.99% of the time), their will be more VideoFormats than network
interfaces on a device( Service Provider or Service User).
So the new implementation is to initialize all channels each time a new
videoFormat is detected, as long as all channels are not mapped to a
videoFormat. I'm explaining myself:
If there is two network interfaces (whihch is usually the minimum that a
computer has, because their is the loopback, and a classic NI: eth0 or
ens2p0): their is potentially two video chanels. So two entries in the
table channelTable. Then, for the first 2 videoFormats detected on the
device (for the first two streams initialized) we will map this streams
to the two channels associated to two different multicast addresses.
Then, all new videoFormat will be added into videoFormatTable, but will
not be added to the channelTable (no mapping will be done between any
channel and those new videoFormats). So if the user want to start to
stream them, he will need first to change the mapping between a
videoFormat and a Channel by setting the channelVideoFormatIndex
parameter to a new value in the channelTable and the set the
channelStatus to 1 (start).