Add the notion of Primary and secondaries interfaces
In VIVOE Primaru interface is the interface the handle SNMP request.
Others are ONLY used to stream, which means they are only used to send
data on a multicast IP address. The primary interface can both stream,
and handle SNMP requests. This interface also dispatch the stream on the
different network interfaces as requested by the user through SNMP. Now
the primary interface is defined as the DefaultReceiveIP, and is defined
as primary by the user in the configuration file: it is the first
interface of the list of ethernet interfaces defined by the user.