Add the possibility to assgined a new videoFormat to a channel
Before, one channel was created for one video format, then I changed
that to build one channel for one ethernetInterface, but the first
implementation was the good one. So I revert that, and we are back to
the first implementation. However, the first implementation does not
give the possibility to update the videoFormat assigned to a channel.
This commit gives the possibility to change the videoFormat associated
to a channel.