Add a way to choose the sink for type detection
To detect the video Caps of our stream in the pipeline, we use the
typefinder element of Gstreamer, that is connected to the signal "have
type". This element, typefind, is connected to an input element of our
choice and to an output element, which shall be a sink. Before, this
sink was always a fakesink, now, it can be any sink that we want in
particular udpsink.