pipeline.c: Add a parser for MPEG-4 videos
Before, the typedind (in detect.c) did not work to detect the video Caps
of MPEG-4 video payloaded by the RTP-MPEG-4 payloader. So the solution
was to add a MPEG-4 parser before the rtpmp4vpay so that it parses the
video stream, and gives us more information than the simple caps detected
with typedfind.