Delete videoFormatIndex member in stream_data structure
Before, into the steam_data, was saved the index of the
VideoFormat associated to the channel that contains those stream datas.
This is superfluous information as this information is already stored in
the parameter: channelVideoFormat of the channelTable. So I have deleted
it. However, it implied to modify the handler of signals SIGINT and
SIGTERM, as well as the function, start_stream, stop_stream, and
delete_stream. So all these function and their callings have been
modify.