stream.c: set pipeline in NULL state before deleting it
This correct a bug introduce in previous commit. When Gstreamer delete a
pipeline, it deletes all elements contained into it. But to delete them,
the pipeline need to be in NULL state. Before last commit this was done
in stop_streaming, and both of those functions were called
simultaneously. Now the delete_streaming function sets the pipeline into
NULL state before deleting it.