stream_registration.c: coorect a bug to fill videoFormatInterlaced
Before, if no field "interlace-mode" was found in the video caps of the
stream. This was done in function fill_entry. This is not the
behaviour that we want. The problem is that, after payloading the RAW
video, the caps of the stream changed, and there is no more
"interlace-mode" present. So this parameter of the entry
videoFormat_entry was always set to none even if it had been set to
"interlaced" orr "progressive" before. This have been corrected.