Correct an error in the build of SDP messages
Indeed, before this commit session_id and session_version were two
DIFFERENT integers generated randomly. However, the norm specifies that
they should indeed be generated randomly but that they should also be
the same, more precisely session version should be the same as
session ID. This is now the case.