Create sap.h and sap.c files for SAP announcement
The VIVOE norm specifies that the SDP files used to decode a stream
should be send using the SAP protocol, for Session Announcement
Protocol, which is a UDP based protocol, running on port 9875. The files
sap.[ch] will be used to open the UDP socket on port 9875, create the
UDP payload, which will contain a SAP header, and a SAP payload (which
contain the SDP itself).