ethetnetIfTable.c reverse the list of entries in ethernetIfTable
Before the entries in the ethernetIfTable were added at the head of the
list: each time a new entry was added to the list, it become the head.
However, this is not in adequation with was we want. Because it means
that the primary interace will be the last element of the list of
entries in the table. This commit inverse the way to intiate the list:
the element are added at the end, and no longer at the head as before.