ethernetIfTable Make the table accessible in other files
Before the ethenetIfTable was only accesible from ethernetIfTable.c
because, it was not needed in other part of the code, as the IP (or MAC)
addresses were not accessed. The channel Table need the primary
interface name to build the multicast address but there are no need for
the actual IP address (at least in the way the code has been thinking).
Anyway, the SDP files need the IP address, so we may get it by using the
fonction of ethernetIfTable_init but as the value is already saved in
the global table... We don't want to do the work twice!
So making the ethernetIfTable available in other part of the code, makes
it possible to create the SAP header without perform any ioctl called to
retrieve the IP address of the primary interface.