Create a function to build SAP messages
Now, we are able to build a SAP message entirely, by building
successively the header and the payload. To build the header, we have to
access the ethernetIfTable to extract the IP of the primary interface.
Then, we create a SDP message once, retrieve its session_version number,
use it to build the Message Hash Identifier. Then we build the SAP
payload which is just a textual representation of the SDP message
(thanks to gstreamer API each lines are already terminated by "\r\n"
thus, in hexadecimal representation: 0x0D0A, which is the behaviour we
want in the VIVOE norm). Finnaly we concatenate the header and the
payload into one single byte array which is our SAP message that will be
used as the UDP payload for the UDP datagram.