Send SAP/SDP over UDP every second when start signal received on channel
Now, each time the channelStatus is sent to one by a SNMP request from
the manager, not only does openvivoe start the stream, but it also start
to send the SAP/SDP announcements associated to this stream every
second. I have captured those packets with wireshark, and they are
correclty build.
A new features is that the SDP message and then the SAP payload are
built at the same time we register the channel. So they are only build
once, then when the channel Status is set to start, asynchronously, the
job added to the main loop and called periodicaly every second just
performs the sending of the UDP packet, that is already built, on the
well known multicast address "224.2.127.254" and well known port 9875.