Send multiple announcements if multiple channels are started
With the previous implementation, only one anouncement was send even if
multiple channels were in "start" mode. However, we want to be able to
send as many SAP/SDP announcement as the number of channels in "start"
mode and that every seconds (on average).
Thus two implementation can be done:

The first one, we build an array of sap_data, which registers in each
cell the sap_data associated to a particular stream: for example:
sap_datas[0] --> sap_data of channelIndex.1
sap_datas[1] --> sap_data of channelIndex.2
sap_datas[2] --> sap_data of channelIndex.3
etc...
And then every seconds, we called a function that send the UDP datagram
if channelStatus is on start for all sap_data registered in the array.
However, this implementation does not take into account the fact that
channels can be launch asynchronously and with different SAP annoucement
intervals. For example, channel 1 could be launch with a SAP interval of
1 second, when channel 2 could be launched with a SAP interval of 10
second. It means that we must keep a track of previous calls and for each
channel test if the current call is a multiple our the
channelSAPinterval, and then choose to send or not the SAP/SDP
announcement. This can be tricky, as we also must be able to reset this
period if a request to set channelStatus to start is received.
The main advantage if this method is that we will implement a global
variable that we be unique to the sap.c file.

Another implementation is to create a member of type sap_data in the
structure channelTable_entry. Thus, each entry in the channelTable is
associated to a particular SAP/SDP message (to a particuler UDP
payload). Then, when a request to set channelStatus to start is received,
we add to the GMainLoop a preiodic call to the function that send the
UDP datagram. The big advantage is that we don't have to worry about the
period to send the SAP/SDP message, as it will be automatically reset
each time a channelStatus is received and that all channels will have
their own timeline.

I have choose to implement the second implementation.