Create the SAP deletion message
Before, when a channel was set to stop, it stopped to send SAP/SDP
annoucement message, but it did not send the SAP deletion message
as it is specified in the norm. Now, the channel send a deletion message
before extracting the job, that call the send_SAP function periodically,
from the GMainLoop.