Fixes in the way to build SAP announcements
SAP announcements were created when the channel was created
however, it leads to a bug when the manager performs the following
sequence:
1 - Set channelStatus to start
2 - Set channelStatus to stop
3 - Set channelStatus to start
Indeed, in step 2, the SAP message built is a deletion message, but
this deletion message is registered in the sap_data member of the
corresponding channelTable_entry. So the fix constits to build the SAP
announcement message each time the channelStatus is set to start and no
only when the channel_entry is created.