sdp.c: delete a errorneous log message
There was a test done on the IP used to fill the file <address> of
origin in the sdp message. It tested if the IP used was a multicast IP,
but as explained in the previous commit, this was not the IP we want,
the IP we want into that field should not be a multicast address. So
the error message have been deleted.