Set period of SAP to channelSAPMessageInterval
Before the only period to send the SAP messages was 1 second, and we did
not take into account the fact that the user can change this period.
Now the period used for sending SAP messages is 1 second by default but
it can be changed by the user with a write SNMP request to the
parameter: channelSAPMessageInterval.