channelTable.c: add a way to create an empty entry
Service User need to have a channel. Indeed, to start displaying a
stream, the service User should receive from the manager a SNMP request
that specify the IP it should start to display and that set the
channelStatus to "start" (1). So, serviceUser should always have a
channel available, which should be an empty channel: with no values for
streaming parameters or for the receiveIp, as well as for the SAP
message interval. So we need a function that create an entry in the
channelTable with no parameters.