stream_registreation: change the way to defined channelIndex
When creating new channels into stream_registration.c, the channelIndex
used was the same index as the VideoFormatIndex associated to the
currently created channel. However, in the case that the device is both
a serviceProvider, and a ServiceUser, at least one emtpy channel has
been created for the SericeUser, so the index to use is not the
videoFormatIndex of the videoFormat associated to the channel but should
be ChannelNumber + 1. This has been corrected.