channelTable.c: create a handler for channelStatus
ChannelStatus is a key member of the structure channelTable_entry.
Indeed, it has specific and different behaviour depending on whether the
channel is a videoChannel or a serviceUser. In the first case it should
build and send SAP/SDP messages, in the second case it should open a
socket and listen for datagrams from 224.2.127.254 port 9875. A handler
that calls the specific functions has been created to simplify the code.