main: differenciate SU and SP, channelTalbe:correct bug channel creation
Main.c: add the differentiation between ServiceUser and ServiceProvider.
In the second case we will intiate streams.
However, as we can launch only one instance of the program (because the
MIB can only be handled by one program), we will use a device stated as
both a SP and a SU.
channelTable: when creating empty channels for the SU, I forgot to
increase the number of channel already present, so I corrected that.