Create a global var for the socket addr for SAP/SDP annoucements
Before, each channel creates a socket_addr (contained into a socket_in)
which contains a port number as well as a struct in_addr which was
always 9875 and 224.2.127.254 for the VIVOE norm. As the socket address
to send the SAP/SPD announcement is the same for all channel, it makes
more sense to save it into a global constant variable, local to the sap.c
file. So the addr_info member of sap_data has been removed, which saves
a lot of memory (sizeof(arrd_info) * channelNumber).