sap.c: remove socket parameters from sap_data.
Previously sap_datas was a structure representing all the data we needed
to send the SAP/SDP announcement of a channel, without needed to rebuild
the SAP/SDP annoucement, every second. However, this structure was
composed of 4 members:
_ a sock_addr_in: a structure to represent the socket address on which
to send the announcement.
_ a int udp_fd the file descriptor of the openned socket for the channel.
_ a char* payload, the SAP message.
_ a int payload, the length of the above member.
So for each channel a socket was openned on the same socket address
("224.2.127.254:9875") however, as this socket address is the same for
all the channels we only need to open one.
This commit creates and opens a unique global socket, that is used by
every channel to send their SAP/SDP announcement, and the SAP message is
still only build once, and register into the sap_data member of each
channel.