hanlder.c: change the way to save old values
The previous implementation set up to support UNDO did not take into
account the fact that the handler is now share between all parameters of
the MIB of the same type. For example the value of old_integer was reset
to NULL for each request->mode. A quick and easy way to prevent from
this was to create two global variables for RW parameters: an integer
and a string. When a SET request is performed, the value of the
parameter is saved into old[integer-string]. If the SET_ACTION is
performed without errors, we do not erase the content of
old[integer-string] but it does not matter as  it will be over written
during the next SET request during the SET_ACTION. If an error occur
during SET_ACTION, SET_UNDO will restore the content of the value as the
normal behaivour.