ServiceUser: get SAP/SDP annoucement from SP
We are able to send SAP/SDP announcements when the channelType was
videoChannel. Now we are also able to read the SAP/SDP announcements
received on the socket: 224.2.127.254:9875. Before, we were not able to
do so because the function used was receivefrom, using the specific
multicast address. It was working sometimes, but after receiving
several packets, it stopped working. Using read() to read on the opened
socket is way more efficient (accordingly to the tests done).