SAP/SDP: Retrieve video caps from SAP/SDP message
We were able to read for the socket the SAP/SDP announcement received on
a channel which is a serviceUser. Functions have been added to:
_ extract from the UDP datagrams the SDP message (payload of SAP
message).
_ Get the media field of the SDP message, and build video caps from this
field.
_ In gstreamer-1.8 a function will be added to get the video caps from
SDP media. However as we are using the version 1.6.3 I just got the
source code, and implemented it in the code of openvivoe.