create_pipeline becames create_pipeline_videoChannel
The function create_pipeline used in stream.c to create the
corresponding pipeline has been rename
create_pipeline_videoChannel() because it corresponds to the creation of
a pipeline for a video channel (with an rtp_payloader and a udpsink).
But now, we also want to build pipeline for serviceUser with udpsrc and
RTP depayloader.