Service User get multicast group from MIB
The VIVOE norm specifies that, to start a stream the manager should send
a SNMP request to the ServiceUser first, specifying the IP multicast it
should joind to get the stream, and a second SNMP request to set the
channelStatus to start. The SU should join the SAP/SDP multicast group,
get the right announcement, and display the stream.
Before, for test purposes, I had hardly written the multicast stream
group into the code. However, it is not the behaviour wanted. Now the SU
will get the multicast stream IP registered in channelReceiveIP to know
which stream to display.